+++ game/src/dlls/objects/671_ARWBombColl/ARWBombColl.c
@@ -138,8 +138,8 @@ void ARWBombColl_update(GameObject* obj)
             ObjHits_SetHitVolumeSlot(&obj->anim, ARWBOMBCOLL_HIT_VOLUME_SLOT, 0, 0);
             if (flags->shotOpen != 0)
             {
-                if ((u32)((ObjHitsPriorityState*)objAnim->hitReactState)->lastHitObject != 0 &&
-                    (u32)((ObjHitsPriorityState*)objAnim->hitReactState)->lastHitObject == (u32)getArwing())
+                if (((ObjHitsPriorityState*)objAnim->hitReactState)->lastHitObject != 0 &&
+                    ((ObjHitsPriorityState*)objAnim->hitReactState)->lastHitObject == (uintptr_t)getArwing())
                 {
                     arwarwing_addScore(arw, 0x19);
                     flags->collected = 1;
@@ -158,8 +158,8 @@ void ARWBombColl_update(GameObject* obj)
                     Obj_SetActiveModelIndex(obj, 1);
                     spawnExplosion(obj, sExplosionScale, 1, 0, 0, 0, 0, 0, 2);
                 }
-                if ((u32)((ObjHitsPriorityState*)objAnim->hitReactState)->lastHitObject != 0 &&
-                    (u32)((ObjHitsPriorityState*)objAnim->hitReactState)->lastHitObject == (u32)getArwing())
+                if (((ObjHitsPriorityState*)objAnim->hitReactState)->lastHitObject != 0 &&
+                    ((ObjHitsPriorityState*)objAnim->hitReactState)->lastHitObject == (uintptr_t)getArwing())
                 {
                     obj->anim.flags |= OBJANIM_FLAG_HIDDEN;
                     ObjHits_DisableObject(obj);
